requirements.txt
requests
python-dotenv
tabulate
Flask
flask_caching
redis
gunicorn==21.2.0
